+++ db/migrate/20250302123208_create_line_items.rb
@@ -3,11 +3,11 @@ class CreateLineItems < ActiveRecord::Migration[7.0]
     create_table :line_items do |t|
       t.references :invoice, null: false, foreign_key: true
       t.string :description, null: false
-      t.decimal :quantity, null: false, precision: 10, scale: 2
-      t.decimal :unit_price, null: false, precision: 10, scale: 2
-      t.decimal :total, null: false, precision: 10, scale: 2
+      t.float :quantity, null: false, precision: 10, scale: 2
+      t.float :unit_price, null: false, precision: 10, scale: 2
+      t.float :total, null: false, precision: 10, scale: 2
 
       t.timestamps
     end
-  end
+  end 
 end
